+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Support for bitstring pattern matching
   * Support for bitstrings in for comprehensions
   * Support for functions with catch, after, else clauses
+  * Support for `with` special form 
 
 # v0.15.2
 * Enhancements